+++ src/main/java/duke/dukeexceptions/DukeException.java
@@ -1,3 +1,5 @@
+package duke.dukeexceptions;
+
 public class DukeException extends Exception {
     public DukeException(String s) {
         super(s);
