+++ src/main/java/duke/dukeexceptions/Missing.java
@@ -1,3 +1,5 @@
+package duke.dukeexceptions;
+
 public class Missing extends DukeException {
     public Missing(String s) {
         super(s);
